==> 1.feature_engineering.ipynb <==
# Import packages
from collections import Counter
import csv
from datetime import datetime
import matplotlib.pyplot as plt
import numpy as np
import pandas as pd
import re
import spacy
from spacy.lang.en.stop_words import STOP_WORDS
from spacytextblob.spacytextblob import SpacyTextBlob
import time
import warnings
warnings.filterwarnings("ignore")
%matplotlib inline
plt.rcParams.update({"font.size": 12})

# 1. Load Dataset

submissions_csv = "data/submissions.csv"
comments_csv = "data/comments_pratik.csv"
feature_comments_csv = "data/feature_comments_rakib_new.csv"

date_fmt = lambda x: datetime.utcfromtimestamp(int(x)).strftime("%Y-%m-%d")

## 1.1. Load submissions data

# Load submissions data
df_submissions = pd.read_csv(submissions_csv, delimiter=";", parse_dates=["created_utc"],
                 date_parser=date_fmt, encoding="utf-8")
df_submissions.info()

df_submissions[["id","created_utc", "author","link_flair_text","title","selftext","num_comments","score",
                "upvote_ratio","removed_by_category"]].tail(1000)[40:45]

## 1.2. Load comments data

df_comments = pd.read_csv(comments_csv, delimiter=";", parse_dates=["created_utc"],
                date_parser=date_fmt, encoding="utf-8")
df_comments.info()

df_comments[["id","created_utc","author","body","score","link_id"]].head(20)[8:13]

## 1.3. Load S&P 500 tickers

# Load S&P 500 tickers
with open("data/sp500-symbol-list.txt") as tickers:
    sp500 = [ticker.rstrip() for ticker in tickers]

## 1.4. Load non S&P 500 tickers

# Load non S&P 500 tickers
nasdaq_stocks = pd.read_csv("data/nasdaq_stocks.csv", sep=",")
nasdaq_stocks_symbols = nasdaq_stocks["Symbol"].tolist()
non_sp500 = list(set(nasdaq_stocks_symbols) - set(sp500))

## 1.5. Load lingos

# Walstreetbets lingo ref: https://economictimes.indiatimes.com/markets/stocks/news/from-stonks-to-bagholder-deciphering-reddits-trading-lingo
lingo = ["YOLO", "BAGHOLDER", "TENDIES", "DD", "GO", "BUY", "BUYING", "buying", "go", "holding", "high", 
         "moon", "squeeze"]

# 2. Preprocess Data

## 2.1. Filter removed submissions

Remove submissions that are deleted by author or moderator or Reddit filter.
- **anti_evil_ops, author, deleted**: Post was removed by author.
- **automod_filtered**: Post is awaiting moderator approval.
- **moderator**: Post was removed by moderator.
- **reddit**: Post was removed by Reddit's spam filters.

df_submissions = df_submissions[df_submissions["removed_by_category"].isnull()]
df_submissions.drop("removed_by_category", axis=1, inplace=True)
df_submissions.head()

## 2.2. Filter removed comments

Remove comments associated with removed submissions. Only keep comments from non-removed submissions.

df_comments = df_comments[df_comments["author"] != "[deleted]"]
df_comments = df_comments[~df_comments["created_utc"].isna()]

df_comments["parent_id"] = df_comments["link_id"].apply(lambda x: x.split("_")[1])
df_comments.head()

# Do inner join which will remove the comments whose id is not in submissions dataframe
df_comments = pd.merge(df_comments, df_submissions, left_on="parent_id", right_on="id", how="inner")
df_comments.head()

# Create new comments dataframe
df_comments = df_comments[["author_x", "body", "created_utc_x", "id_x", "score_x", "parent_id"]]
df_comments.head()

# 3. Feature Engineering

## 3.1 Submissions

### 3.1.1. Calculate sentiment per day

nlp = spacy.blank("en")
if "spacytextblob" not in nlp.pipe_names:
    nlp.add_pipe("spacytextblob")
    nlp.rename_pipe("spacytextblob", "sentiment")

if "emoji" not in nlp.pipe_names:
    nlp.add_pipe("emoji")
    
print(nlp.pipe_names)

df_submissions["title_polarity"] = 0.0
df_submissions["title_subjectivity"] = 0.0
df_submissions["body_polarity"] = 0.0
df_submissions["body_subjectivity"] = 0.0
df_submissions["emojis"] = ""

def get_submission_sentiment(title, body):
    nlp_title = nlp(str(title))
    nlp_body = nlp(str(body))
    emojis = nlp_title._.emoji + nlp_body._.emoji
    emojis = [emoji[0] for emoji in emojis]
    return pd.Series([nlp_title._.polarity, nlp_title._.subjectivity,
                      nlp_body._.polarity, nlp_body._.subjectivity, emojis])

df_submissions[["title_polarity", "title_subjectivity", "body_polarity", "body_subjectivity", "emojis"]] = df_submissions.apply(lambda row: get_submission_sentiment(row["title"], row["selftext"]), axis=1)
df_submissions.head()

# # Group by date
df_submissions_title_subjectivity = df_submissions.groupby(["created_utc"])["title_subjectivity"].sum().reset_index()

df_submissions_body_polarity = df_submissions.groupby(["created_utc"])["body_polarity"].sum().reset_index()
df_submissions_body_subs_title_polarity = df_submissions.groupby(["created_utc"])["title_polarity"].sum().reset_index()
df_submissionjectivity = df_submissions.groupby(["created_utc"])["body_subjectivity"].sum().reset_index()

### 3.1.2. Calculate emojis per day

def sanitize_emojis(data):
    return data.replace("[", "").replace("]", "").replace(",", "").replace("'", "").replace(" ", "")

df_submissions_emojis = df_submissions.groupby(["created_utc"]).apply(lambda x: str(list(x["emojis"]))).to_frame("emojis").reset_index()
df_submissions_emojis["emojis"] = df_submissions_emojis["emojis"].apply(sanitize_emojis)
df_submissions_emojis["emojis"] = df_submissions_emojis.apply(lambda x: dict(Counter(x["emojis"])), axis=1)

### 3.1.3. Calculate mentioned stocks and lingos count per day

# Function to find stocks in a text given stock list
regex = re.compile("[^a-zA-Z ]")
sp500_set = set(sp500)
non_sp500_set = set(non_sp500)
lingo_set = set(lingo)

def calculate_mentioned_stocks(title, body=""):
    content = regex.sub("", str(title) + " " + str(body)).split(" ")
    content = set(content)
    words_sp500 = str(list(sp500_set & content))
    words_non_sp500 = str(list(non_sp500_set & content))
    
    if len(words_sp500) > 2:
        lingos_sp500 = str(list(lingo_set & content))
    else:
        lingos_sp500 = str([])
    
    if len(words_non_sp500) > 2:
        lingos_non_sp500 = str(list(lingo_set & content))
    else:
        lingos_non_sp500 = str([])
    
    return pd.Series([words_sp500, words_non_sp500, lingos_sp500, lingos_non_sp500])

df_submissions[["sp500_stocks", "non_sp500_stocks", "sp500_lingos", "non_sp500_lingos"]] = df_submissions.apply(lambda x: calculate_mentioned_stocks(x["title"], x["selftext"]), axis=1)
df_submissions.head()

# # Group by date
df_submissions_sp500_stocks = df_submissions.groupby(["created_utc"])["sp500_stocks"].apply(",".join).reset_index()
df_submissions_non_sp500_stocks = df_submissions.groupby(["created_utc"])["non_sp500_stocks"].apply(",".join).reset_index()

df_submissions_sp500_lingos = df_submissions.groupby(["created_utc"])["sp500_lingos"].apply(",".join).reset_index()
df_submissions_non_sp500_lingos = df_submissions.groupby(["created_utc"])["non_sp500_lingos"].apply(",".join).reset_index()

def sanitize_column(column):
    column = column.replace("[],", "").replace("[]", "").replace("],[", ",")
    return column

# # Sanitize data
df_submissions_sp500_stocks["sp500_stocks"] = df_submissions_sp500_stocks["sp500_stocks"].apply(sanitize_column)
df_submissions_non_sp500_stocks["non_sp500_stocks"] = df_submissions_non_sp500_stocks["non_sp500_stocks"].apply(sanitize_column)

df_submissions_sp500_lingos["sp500_lingos"] = df_submissions_sp500_lingos["sp500_lingos"].apply(sanitize_column)
df_submissions_non_sp500_lingos["non_sp500_lingos"] = df_submissions_non_sp500_lingos["non_sp500_lingos"].apply(sanitize_column)

# # Count stocks and lingos
df_submissions_sp500_stocks["sp500_stocks"] = df_submissions_sp500_stocks["sp500_stocks"].str.strip("[]").str.split(",")
df_submissions_sp500_stocks["sp500_stocks"] = df_submissions_sp500_stocks.apply(lambda x: dict(Counter(x['sp500_stocks'])), axis=1)

df_submissions_non_sp500_stocks["non_sp500_stocks"] = df_submissions_non_sp500_stocks["non_sp500_stocks"].str.strip("[]").str.split(",")
df_submissions_non_sp500_stocks["non_sp500_stocks"] = df_submissions_non_sp500_stocks.apply(lambda x: dict(Counter(x["non_sp500_stocks"])), axis=1)

df_submissions_sp500_lingos["sp500_lingos"] = df_submissions_sp500_lingos["sp500_lingos"].str.strip("[]").str.split(",")
df_submissions_sp500_lingos["sp500_lingos"] = df_submissions_sp500_lingos.apply(lambda x: dict(Counter(x["sp500_lingos"])), axis=1)

df_submissions_non_sp500_lingos["non_sp500_lingos"] = df_submissions_non_sp500_lingos["non_sp500_lingos"].str.strip("[]").str.split(",")
df_submissions_non_sp500_lingos["non_sp500_lingos"] = df_submissions_non_sp500_lingos.apply(lambda x: dict(Counter(x["non_sp500_lingos"])), axis=1)

### 3.1.4. Count per day

df_submissions_count = df_submissions.groupby(["created_utc"])["title"].count().reset_index()
df_submissions_count.rename(columns={"title": "count"}, inplace=True)

### 3.1.5. Calculate total scores per day

df_submissions_score = df_submissions.groupby(["created_utc"])["score"].sum().reset_index()

### 3.1.6. Calculate total upvote ratio per day

df_submissions_upvote_ratio = df_submissions.groupby(["created_utc"])["upvote_ratio"].sum().reset_index()

### 3.1.7. Calculate flair category counts per day

def sanitize_category(df):
    category_dict = dict(Counter(df["category"]))
    if np.nan in category_dict.keys():
        category_dict["others"] = category_dict.pop(np.nan)
    return category_dict

df_submissions_category = df_submissions.groupby(["created_utc"]).apply(lambda x: list(x["link_flair_text"])).to_frame("category").reset_index()
df_submissions_category["category"] = df_submissions_category.apply(lambda x: sanitize_category(x), axis=1)

### 3.1.8. Create the final submissions dataframe

df_submissions_daily = pd.concat([df_submissions_sp500_stocks["created_utc"], 
                                  df_submissions_title_polarity["title_polarity"], 
                                  df_submissions_title_subjectivity["title_subjectivity"], 
                                  df_submissions_body_polarity["body_polarity"], 
                                  df_submissions_body_subjectivity["body_subjectivity"], 
                                  df_submissions_count["count"],
                                  df_submissions_score["score"],
                                  df_submissions_upvote_ratio["upvote_ratio"],
                                  df_submissions_emojis["emojis"],
                                  df_submissions_category["category"],
                                  df_submissions_sp500_stocks["sp500_stocks"], 
                                  df_submissions_non_sp500_stocks["non_sp500_stocks"],
                                  df_submissions_sp500_lingos["sp500_lingos"], 
                                  df_submissions_non_sp500_lingos["non_sp500_lingos"]], axis=1)

df_submissions_daily.tail()

### 3.1.9. Save the submissions dataframe

df_submissions_daily.to_csv("data/feature_submissions.csv", sep=";", index=False)

## 3.2. Comments

### 3.2.1 Calculate sentiment per day

df_comments["body_polarity"] = 0.0
df_comments["body_subjectivity"] = 0.0
df_comments["emojis"] = ""

def get_comment_sentiment(body):
    nlp_body = nlp(str(body))
    emojis = nlp_body._.emoji
    emojis = [emoji[0] for emoji in emojis]
    return pd.Series([nlp_body._.polarity, nlp_body._.subjectivity, emojis])

df_comments[["body_polarity", "body_subjectivity", "emojis"]] = df_comments.apply(lambda row: get_comment_sentiment(row["body"]), axis=1)
df_comments.head()

# Group by date
df_comments["body_polarity"] = df_comments["body_polarity"].fillna(0)
df_comments["body_polarity"] = df_comments["body_polarity"].astype(float)
df_comments_body_polarity = df_comments.groupby(["created_utc_x"])["body_polarity"].sum().reset_index()

df_comments["body_subjectivity"] = df_comments["body_subjectivity"].fillna(0)
df_comments["body_subjectivity"] = df_comments["body_subjectivity"].astype(float)
df_comments_body_subjectivity = df_comments.groupby(["created_utc_x"])["body_subjectivity"].sum().reset_index()

### 3.2.2. Calculate emojis per day

df_comments_emojis = df_comments.groupby(["created_utc_x"]).apply(lambda x: str(list(x["emojis"]))).to_frame("emojis").reset_index()
df_comments_emojis["emojis"] = df_comments_emojis["emojis"].apply(sanitize_emojis)
df_comments_emojis["emojis"] = df_comments_emojis.apply(lambda x: dict(Counter(x["emojis"])), axis=1)

### 3.2.2. Calculate mentioned stocks and lingos count per day

df_comments[["sp500_stocks", "non_sp500_stocks", "sp500_lingos", "non_sp500_lingos"]] = df_comments.apply(lambda x: calculate_mentioned_stocks(x["body"]), axis=1)
df_comments.head()

# Group by date
df_comments_sp500_stocks = df_comments.groupby(["created_utc_x"])["sp500_stocks"].apply(",".join).reset_index()
df_comments_non_sp500_stocks = df_comments.groupby(["created_utc_x"])["non_sp500_stocks"].apply(",".join).reset_index()

df_comments_sp500_lingos = df_comments.groupby(["created_utc_x"])["sp500_lingos"].apply(",".join).reset_index()
df_comments_non_sp500_lingos = df_comments.groupby(["created_utc_x"])["non_sp500_lingos"].apply(",".join).reset_index()

# Sanitize data
df_comments_sp500_stocks["sp500_stocks"] = df_comments_sp500_stocks["sp500_stocks"].apply(sanitize_column)
df_comments_non_sp500_stocks["non_sp500_stocks"] = df_comments_non_sp500_stocks["non_sp500_stocks"].apply(sanitize_column)

df_comments_sp500_lingos["sp500_lingos"] = df_comments_sp500_lingos["sp500_lingos"].apply(sanitize_column)
df_comments_non_sp500_lingos["non_sp500_lingos"] = df_comments_non_sp500_lingos["non_sp500_lingos"].apply(sanitize_column)

# Count stocks and lingos
df_comments_sp500_stocks["sp500_stocks"] = df_comments_sp500_stocks["sp500_stocks"].str.strip("[]").str.split(",")
df_comments_sp500_stocks["sp500_stocks"] = df_comments_sp500_stocks.apply(lambda x: dict(Counter(x['sp500_stocks'])), axis=1)

df_comments_non_sp500_stocks["non_sp500_stocks"] = df_comments_non_sp500_stocks["non_sp500_stocks"].str.strip("[]").str.split(",")
df_comments_non_sp500_stocks["non_sp500_stocks"] = df_comments_non_sp500_stocks.apply(lambda x: dict(Counter(x["non_sp500_stocks"])), axis=1)

df_comments_sp500_lingos["sp500_lingos"] = df_comments_sp500_lingos["sp500_lingos"].str.strip("[]").str.split(",")
df_comments_sp500_lingos["sp500_lingos"] = df_comments_sp500_lingos.apply(lambda x: dict(Counter(x["sp500_lingos"])), axis=1)

df_comments_non_sp500_lingos["non_sp500_lingos"] = df_comments_non_sp500_lingos["non_sp500_lingos"].str.strip("[]").str.split(",")
df_comments_non_sp500_lingos["non_sp500_lingos"] = df_comments_non_sp500_lingos.apply(lambda x: dict(Counter(x["non_sp500_lingos"])), axis=1)

### 3.2.3. Count  per day

df_comments_count = df_comments.groupby(["created_utc_x"])["body"].count().reset_index()
df_comments_count.rename(columns={"body": "count"}, inplace=True)

### 3.2.4. Calculate total scores per day

df_comments["score_x"] = df_comments["score_x"].fillna(0)
df_comments["score_x"] = df_comments["score_x"].astype(int)
df_comments_score = df_comments.groupby(["created_utc_x"])["score_x"].sum().reset_index()

### 3.2.5. Create the final comments dataframe

df_comments_daily = pd.concat([df_comments_sp500_stocks["created_utc_x"], 
                               df_comments_body_polarity["body_polarity"], 
                               df_comments_body_subjectivity["body_subjectivity"], 
                               df_comments_count["count"],
                               df_comments_score["score_x"],
                               df_comments_emojis["emojis"],
                               df_comments_sp500_stocks["sp500_stocks"], 
                               df_comments_non_sp500_stocks["non_sp500_stocks"],
                               df_comments_sp500_lingos["sp500_lingos"], 
                               df_comments_non_sp500_lingos["non_sp500_lingos"]], axis=1)

df_comments_daily.head()

### 3.2.6. Save the comments dataframe

df_comments_daily.to_csv(feature_comments_csv, sep=";", index=False)

# df_comments.drop([4929153, 4929185, 4929199, 4929235, 4929244, 5015671, 5015833, 5016105, 5021356], inplace=True)
