==> 2.feature_creation.ipynb <==
# Import packages
import ast
from collections import Counter, defaultdict
import numpy as np
import pandas as pd

# Combine Comments Dataframe

# Load comment features
df_comments_pratik = pd.read_csv("data/feature_comments_pratik.csv", delimiter=";", encoding="utf-8")
df_comments_rakib = pd.read_csv("data/feature_comments_rakib.csv", delimiter=";", encoding="utf-8")
df_comments_dami = pd.read_csv("data/feature_comments_dami.csv", delimiter=";", encoding="utf-8")

# Merge three comment dataframes to create a single comment dataframe
df_comments_combined = pd.concat([df_comments_rakib, df_comments_pratik, df_comments_dami], axis=0, ignore_index=True)

def agg_counter_objs(rows):
    new_counter_objs = []
    for row in rows:
        row_dict = ast.literal_eval(row)
        
        def default_val():
            return 0
        
        new_dict = defaultdict(default_val)
        for key in row_dict.keys():
            new_key = key.replace("'","").replace('"',"").replace("]","").replace("\\","").strip()
            new_dict[new_key] += row_dict[key]
        
        if "" in new_dict.keys():
            del new_dict[""]
        
        new_counter_objs.append(Counter(new_dict))
    
    agg_counter_obj = Counter({})
    for obj in new_counter_objs:
        agg_counter_obj += obj
        
    return str(dict(agg_counter_obj))

# Group comment dataframe by date
df_comments_grouped = df_comments_combined.groupby("created_utc_x")
df_comments = df_comments_grouped["body_polarity"].sum().reset_index()
df_comments["body_subjectivity"] = df_comments_grouped["body_subjectivity"].sum().reset_index()["body_subjectivity"]
df_comments["count"] = df_comments_grouped["count"].sum().reset_index()["count"]
df_comments["score_x"] = df_comments_grouped["score_x"].sum().reset_index()["score_x"]
df_comments["emojis"] = df_comments_grouped["emojis"].apply(agg_counter_objs).reset_index()["emojis"]
df_comments["sp500_stocks"] = df_comments_grouped["sp500_stocks"].apply(agg_counter_objs).reset_index()["sp500_stocks"]
df_comments["non_sp500_stocks"] = df_comments_grouped["non_sp500_stocks"].apply(agg_counter_objs).reset_index()["non_sp500_stocks"]
df_comments["sp500_lingos"] = df_comments_grouped["sp500_lingos"].apply(agg_counter_objs).reset_index()["sp500_lingos"]
df_comments["non_sp500_lingos"] = df_comments_grouped["non_sp500_lingos"].apply(agg_counter_objs).reset_index()["non_sp500_lingos"]

df_comments.columns = ["created_utc", "body_polarity", "body_subjectivity", "count", "score", "emojis",
                      "sp500_stocks", "non_sp500_stocks", "sp500_lingos", "non_sp500_lingos"]
df_comments.head()

# Save the combined comments dataframe
df_comments.to_csv("data/feature_comments.csv", sep=";", index=False)

# Create Features Dataframe with Submissions and Comments

# Load feature_submissions and feature_comments dataframe
df_submissions = pd.read_csv("data/feature_submissions.csv", delimiter=";", index_col="created_utc", encoding="utf-8")
df_comments = pd.read_csv("data/feature_comments.csv", delimiter=";", index_col="created_utc", encoding="utf-8")

df_submissions[["title_polarity","title_subjectivity","body_polarity","body_subjectivity","emojis","category","sp500_stocks",
                "sp500_lingos","count", "score", "upvote_ratio"]].info()

df_comments[["body_polarity","body_subjectivity","emojis","sp500_stocks","sp500_lingos","count", "score"]].info()

# Create the feature dataframe combining submissions and comments features
df_features = pd.concat([df_submissions[["title_polarity", "title_subjectivity", "body_polarity", "body_subjectivity", 
                                         "count", "score", "upvote_ratio"]], 
                        df_comments[["body_polarity", "body_subjectivity", "count", "score"]]], axis=1)
df_features.columns = ["title_polarity", "title_subjectivity", "submission_body_polarity", "submission_body_subjectivity", 
                      "submission_count", "submission_score", "upvote_ratio", "comment_body_polarity", 
                      "comment_body_subjectivity", "comment_count", "comment_score"]
df_features

df_features.info()

# Combine emojis feature for submissions and comments
df_emojis = pd.concat([df_submissions["emojis"], df_comments["emojis"]], axis=0)
df_emojis = df_emojis.groupby(df_emojis.index).apply(agg_counter_objs).reset_index()
df_emojis = df_emojis["emojis"].apply(lambda x: ast.literal_eval(x))
df_emojis = pd.DataFrame(df_emojis.tolist())
df_emojis["created_utc"] = df_submissions.index
df_emojis.set_index("created_utc", inplace=True)

# Combine sp500_stocks feature for submissions and comments
df_sp500_stocks = pd.concat([df_submissions["sp500_stocks"], df_comments["sp500_stocks"]], axis=0)
df_sp500_stocks = df_sp500_stocks.groupby(df_sp500_stocks.index).apply(agg_counter_objs).reset_index()
df_sp500_stocks = df_sp500_stocks["sp500_stocks"].apply(lambda x: ast.literal_eval(x))
df_sp500_stocks = pd.DataFrame(df_sp500_stocks.tolist())
df_sp500_stocks["created_utc"] = df_submissions.index
df_sp500_stocks.set_index("created_utc", inplace=True)

# Combine non_sp500_stocks feature for submissions and comments
df_non_sp500_stocks = pd.concat([df_submissions["non_sp500_stocks"], df_comments["non_sp500_stocks"]], axis=0)
df_non_sp500_stocks = df_non_sp500_stocks.groupby(df_non_sp500_stocks.index).apply(agg_counter_objs).reset_index()
df_non_sp500_stocks = df_non_sp500_stocks["non_sp500_stocks"].apply(lambda x: ast.literal_eval(x))
df_non_sp500_stocks = pd.DataFrame(df_non_sp500_stocks.tolist())
df_non_sp500_stocks["created_utc"] = df_submissions.index
df_non_sp500_stocks.set_index("created_utc", inplace=True)
df_non_sp500_stocks

# Combine sp500_lingos feature for submissions and comments
df_sp500_lingos = pd.concat([df_submissions["sp500_lingos"], df_comments["sp500_lingos"]], axis=0)
df_sp500_lingos = df_sp500_lingos.groupby(df_sp500_lingos.index).apply(agg_counter_objs).reset_index()
df_sp500_lingos = df_sp500_lingos["sp500_lingos"].apply(lambda x: ast.literal_eval(x))
df_sp500_lingos = pd.DataFrame(df_sp500_lingos.tolist())
df_sp500_lingos["created_utc"] = df_submissions.index
df_sp500_lingos.set_index("created_utc", inplace=True)

# Combine non_sp500_lingos feature for submissions and comments
df_non_sp500_lingos = pd.concat([df_submissions["non_sp500_lingos"], df_comments["non_sp500_lingos"]], axis=0)
df_non_sp500_lingos = df_non_sp500_lingos.groupby(df_non_sp500_lingos.index).apply(agg_counter_objs).reset_index()
df_non_sp500_lingos = df_non_sp500_lingos["non_sp500_lingos"].apply(lambda x: ast.literal_eval(x))
df_non_sp500_lingos = pd.DataFrame(df_non_sp500_lingos.tolist())
df_non_sp500_lingos["created_utc"] = df_submissions.index
df_non_sp500_lingos.set_index("created_utc", inplace=True)

# For submissions category
df_submissions["category"] = df_submissions["category"].apply(lambda x: ast.literal_eval(x))
df_submissions_category = pd.DataFrame(df_submissions["category"].tolist())
df_submissions_category["created_utc"] = df_submissions.index
df_submissions_category.set_index("created_utc", inplace=True)

# Create the feature dataframe that will be feed into machine learning model
df_features = pd.concat([df_features, 
                         df_emojis, 
                         df_submissions_category, 
                         df_sp500_stocks, 
                         #df_non_sp500_stocks, 
                         df_sp500_lingos], axis=1)
                         #df_non_sp500_lingos
df_features

df_features.reset_index(level=0, inplace=True)

df_features.iloc[:5,10:]

df_features.info()

# Save the final features dataframe
df_features.to_csv("data/features_sp500.csv", sep=";", index=False)
