==> 3.data_preparation.ipynb <==
# Import packages
import ast
from collections import Counter
import numpy as np
import pandas as pd

df_submissions = pd.read_csv("data/feature_submissions.csv", delimiter=";", index_col="created_utc", encoding="utf-8")
df_comments = pd.read_csv("data/feature_comments_pratik.csv", delimiter=";", index_col="created_utc_x", encoding="utf-8")

# df_submissions.head()

# df_comments.head()

# Do this with every features of type Counter
df_submissions["emojis"] = df_submissions["emojis"].apply(lambda x: ast.literal_eval(x))
df_submissions_emojis = pd.DataFrame(df_submissions["emojis"].tolist())
df_submissions_emojis["created_utc"] = df_submissions.index
df_submissions_emojis.set_index("created_utc", inplace=True)

df_submissions["category"] = df_submissions["category"].apply(lambda x: ast.literal_eval(x))
df_submissions_category = pd.DataFrame(df_submissions["category"].tolist())
df_submissions_category["created_utc"] = df_submissions.index
df_submissions_category.set_index("created_utc", inplace=True)

df_submissions["sp500_stocks"] = df_submissions["sp500_stocks"].apply(lambda x: ast.literal_eval(x))
df_submissions_sp500_stocks = pd.DataFrame(df_submissions["sp500_stocks"].tolist())
df_submissions_sp500_stocks["created_utc"] = df_submissions.index
df_submissions_sp500_stocks.set_index("created_utc", inplace=True)

df_submissions["non_sp500_stocks"] = df_submissions["non_sp500_stocks"].apply(lambda x: ast.literal_eval(x))
df_submissions_non_sp500_stocks = pd.DataFrame(df_submissions["non_sp500_stocks"].tolist())
df_submissions_non_sp500_stocks["created_utc"] = df_submissions.index
df_submissions_non_sp500_stocks.set_index("created_utc", inplace=True)

df_submissions["sp500_lingos"] = df_submissions["sp500_lingos"].apply(lambda x: ast.literal_eval(x))
df_submissions_sp500_lingos = pd.DataFrame(df_submissions["sp500_lingos"].tolist())
df_submissions_sp500_lingos["created_utc"] = df_submissions.index
df_submissions_sp500_lingos.set_index("created_utc", inplace=True)

df_submissions["non_sp500_lingos"] = df_submissions["non_sp500_lingos"].apply(lambda x: ast.literal_eval(x))
df_submissions_non_sp500_lingos = pd.DataFrame(df_submissions["non_sp500_lingos"].tolist())
df_submissions_non_sp500_lingos["created_utc"] = df_submissions.index
df_submissions_non_sp500_lingos.set_index("created_utc", inplace=True)

# @todo Do the same conversion for comments dataframe
# @todo Maybe we need to combine emojis, sp500 and non sp500 stocks for both submissions and comments

# Create the feature dataframe that will be feed into machine learning model
df_features = pd.concat([df_submissions["title_polarity"]/df_submissions["count"], 
                        df_submissions["body_polarity"]/df_submissions["count"], 
                        df_submissions["score"]/df_submissions["count"], 
                        df_submissions["upvote_ratio"]/df_submissions["count"],
                        df_comments["body_polarity"]/df_comments["count"],
                        df_comments["score_x"]/df_comments["count"]],
                        axis=1)

df_features.columns = ["submission_title_polarity", "submission_body_polarity", "submission_score",
                      "submission_upvote_ratio", "comment_polarity", "comment_score"]

df_features = pd.concat([df_features,
                         df_submissions_category, 
                         df_submissions_emojis,
                         df_submissions_sp500_stocks, 
                         df_submissions_non_sp500_stocks,
                         df_submissions_sp500_lingos, 
                         df_submissions_non_sp500_lingos], axis=1)

column_names = []
for col in df_features.columns:
    name = col.replace("'", "").replace("]", "").replace(" ", "")
    column_names.append(name)
    
df_features.columns = column_names

df_features
